move Buffer class into Protocol module
